@@ -1,48 +1,50 @@
 {
-  "name": "httplaceholder",
+  "name": "gui",
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service lint && vue-cli-service build",
+    "serve": "vue-cli-service serve --port 10001",
+    "build": "vue-cli-service build",
+    "test:unit": "vue-cli-service test:unit",
     "lint": "vue-cli-service lint"
   },
   "dependencies": {
-    "@aspnet/signalr": "^1.1.4",
-    "@mdi/font": "^4.9.95",
-    "axios": "^0.18.0",
-    "core-js": "^3.8.1",
-    "js-yaml": "^3.14.1",
-    "lodash.camelcase": "^4.3.0",
-    "moment": "^2.29.1",
-    "register-service-worker": "^1.7.2",
-    "toastr": "^2.1.4",
-    "typeface-roboto": "0.0.75",
-    "vue": "^2.6.12",
-    "vue-codemirror": "^4.0.6",
-    "vue-router": "^3.4.9",
-    "vue-shortkey": "^3.1.6",
-    "vuetify": "^2.3.21",
-    "vuex": "^3.6.0",
-    "xml-formatter": "2.4.0"
+    "core-js": "^3.6.5",
+    "register-service-worker": "^1.7.1",
+    "vue": "^3.0.0",
+    "vue-router": "^4.0.0-0",
+    "vuex": "^4.0.0-0"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "^3.12.1",
-    "@vue/cli-plugin-eslint": "^3.12.1",
-    "@vue/cli-plugin-pwa": "^3.12.1",
-    "@vue/cli-plugin-router": "~4.3.0",
-    "@vue/cli-plugin-vuex": "~4.3.0",
-    "@vue/cli-service": "^3.12.1",
+    "@microsoft/signalr": "^5.0.10",
+    "@popperjs/core": "^2.10.2",
+    "@vue/cli-plugin-babel": "~4.5.0",
+    "@vue/cli-plugin-eslint": "~4.5.0",
+    "@vue/cli-plugin-pwa": "~4.5.0",
+    "@vue/cli-plugin-router": "~4.5.0",
+    "@vue/cli-plugin-unit-mocha": "~4.5.0",
+    "@vue/cli-plugin-vuex": "~4.5.0",
+    "@vue/cli-service": "~4.5.0",
+    "@vue/compiler-sfc": "^3.0.0",
     "@vue/eslint-config-prettier": "^6.0.0",
+    "@vue/test-utils": "^2.0.0-0",
     "babel-eslint": "^10.1.0",
+    "bootstrap": "^5.1.1",
+    "bootstrap-icons": "^1.5.0",
+    "chai": "^4.1.2",
+    "codemirror": "^5.63.0",
+    "dayjs": "^1.10.7",
     "eslint": "^6.7.2",
-    "eslint-plugin-prettier": "^3.3.0",
-    "eslint-plugin-vue": "^6.2.2",
-    "prettier": "^1.19.1",
-    "sass": "^1.30.0",
-    "sass-loader": "^7.1.0",
-    "vue-cli-plugin-vuetify": "^0.6.3",
-    "vue-template-compiler": "^2.6.12",
-    "vuetify-loader": "^1.6.0"
+    "eslint-plugin-prettier": "^3.3.1",
+    "eslint-plugin-vue": "^7.0.0",
+    "highlight.js": "^11.2.0",
+    "js-yaml": "^4.1.0",
+    "prettier": "^2.2.1",
+    "sass": "^1.26.5",
+    "sass-loader": "^8.0.2",
+    "toastr": "^2.1.4",
+    "typeface-roboto": "^1.1.13",
+    "vue3-slide-up-down": "^1.2.3",
+    "xml-formatter": "^2.4.1"
   }
 }
