@@ -1,9 +1,4 @@
 import "codemirror/lib/codemirror.css";
 import "codemirror/mode/yaml/yaml";
-
-import "./codemirror/style.css";
-import "./codemirror/material-darker.css";
-
-import Vue from "vue";
-import VueCodemirror from "vue-codemirror";
-Vue.use(VueCodemirror);
+import "./codemirror/style.scss";
+import "./codemirror/material-darker.scss";
