@@ -1,7 +1,2 @@
 import "toastr/build/toastr.css";
-import "./toastr/style.css";
-import toastr from "toastr";
-
-toastr.options = {
-  positionClass: "toast-top-full-width"
-};
+import "./toastr/style.scss";
