@@ -1,114 +1,79 @@
-import createInstance from "@/axios/axiosInstanceFactory";
+import { del, get, put, post } from "@/utils/api";
 import yaml from "js-yaml";
 
 const state = () => ({});
 
+const getStub = (stubId) => {
+  return get(`/ph-api/stubs/${stubId}`)
+    .then((response) => Promise.resolve(response))
+    .catch((error) => Promise.reject(error));
+};
+
 const actions = {
   getStubsOverview() {
-    return new Promise((resolve, reject) =>
-      createInstance()
-        .get("ph-api/stubs/overview")
-        .then(response => resolve(response.data))
-        .catch(error => reject(error))
-    );
+    return get("/ph-api/stubs/overview")
+      .then((response) => Promise.resolve(response))
+      .catch((error) => Promise.reject(error));
+  },
+  getStub(_, stubId) {
+    return getStub(stubId);
   },
   getStubs() {
-    return new Promise((resolve, reject) =>
-      createInstance()
-        .get("ph-api/stubs")
-        .then(response => resolve(response.data))
-        .catch(error => reject(error))
-    );
+    return get("/ph-api/stubs")
+      .then((response) => Promise.resolve(response))
+      .catch((error) => Promise.reject(error));
   },
-  /* eslint no-empty-pattern: 0 */
-  getStub({}, payload) {
-    return new Promise((resolve, reject) =>
-      createInstance()
-        .get(`ph-api/stubs/${payload.stubId}`)
-        .then(response => resolve(response.data))
-        .catch(error => reject(error))
-    );
+  async flipEnabled(_, stubId) {
+    const stub = (await getStub(stubId)).stub;
+    stub.enabled = !stub.enabled;
+    await put(`/ph-api/stubs/${stubId}`, stub);
+    return stub.enabled;
   },
-  /* eslint no-empty-pattern: 0 */
-  deleteStub({}, payload) {
-    return new Promise((resolve, reject) =>
-      createInstance()
-        .delete(`ph-api/stubs/${payload.stubId}`)
-        .then(() => resolve())
-        .catch(error => reject(error))
-    );
-  },
-  deleteAllStubs() {
-    return new Promise((resolve, reject) =>
-      createInstance()
-        .delete("ph-api/stubs")
-        .then(() => resolve())
-        .catch(error => reject(error))
-    );
-  },
-  /* eslint no-empty-pattern: 0 */
-  addStubs({}, payload) {
-    return new Promise((resolve, reject) => {
-      let stubsArray;
-      let parsedObject;
-      if (payload.inputIsJson) {
-        parsedObject = payload.input;
-      } else {
-        try {
-          parsedObject = yaml.safeLoad(payload.input);
-        } catch (error) {
-          reject(error.message);
-          return;
-        }
-      }
-
-      if (!Array.isArray(parsedObject)) {
-        stubsArray = [parsedObject];
-      } else {
-        stubsArray = parsedObject;
-      }
-
-      createInstance()
-        .post("ph-api/stubs/multiple", stubsArray)
-        .then(r => r.data)
-        .then(result => resolve(result))
-        .catch(error => reject(error));
-    });
-  },
-  /* eslint no-empty-pattern: 0 */
-  createStubBasedOnRequest({}, payload) {
-    return new Promise((resolve, reject) =>
-      createInstance()
-        .post(`ph-api/requests/${payload.correlationId}/stubs`, {
-          doNotCreateStub:
-            payload.doNotCreateStub !== null ? payload.doNotCreateStub : false
-        })
-        .then(response => resolve(response.data))
-        .catch(error => reject(error))
-    );
-  },
-  async enableStub({}, stubId) {
-    const stub = (await createInstance().get(`ph-api/stubs/${stubId}`)).data
-      .stub;
+  async enableStub(_, stubId) {
+    const stub = (await getStub(stubId)).stub;
     stub.enabled = true;
-    return new Promise((resolve, reject) =>
-      createInstance()
-        .put(`ph-api/stubs/${stubId}`, stub)
-        .then(() => resolve())
-        .catch(error => reject(error))
-    );
+    await put(`/ph-api/stubs/${stubId}`, stub);
+    return stub.enabled;
   },
-  async disableStub({}, stubId) {
-    const stub = (await createInstance().get(`ph-api/stubs/${stubId}`)).data
-      .stub;
+  async disableStub(_, stubId) {
+    const stub = (await getStub(stubId)).stub;
     stub.enabled = false;
-    return new Promise((resolve, reject) =>
-      createInstance()
-        .put(`ph-api/stubs/${stubId}`, stub)
-        .then(() => resolve())
-        .catch(error => reject(error))
-    );
-  }
+    await put(`/ph-api/stubs/${stubId}`, stub);
+    return stub.enabled;
+  },
+  deleteStub(_, stubId) {
+    return del(`/ph-api/stubs/${stubId}`)
+      .then((response) => Promise.resolve(response))
+      .catch((error) => Promise.reject(error));
+  },
+  deleteStubs() {
+    return del("/ph-api/stubs")
+      .then((response) => Promise.resolve(response))
+      .catch((error) => Promise.reject(error));
+  },
+  addStubs(_, input) {
+    const parsedObject = yaml.load(input);
+    const stubsArray = Array.isArray(parsedObject)
+      ? parsedObject
+      : [parsedObject];
+    return post("/ph-api/stubs/multiple", stubsArray)
+      .then((response) => Promise.resolve(response))
+      .catch((error) => Promise.reject(error));
+  },
+  updateStub(_, payload) {
+    const parsedObject = yaml.load(payload.input);
+    return put(`/ph-api/stubs/${payload.stubId}`, parsedObject)
+      .then((response) => Promise.resolve(response))
+      .catch((error) => Promise.reject(error));
+  },
+  createStubBasedOnRequest(_, payload) {
+    return post(`/ph-api/requests/${payload.correlationId}/stubs`, {
+      doNotCreateStub:
+        payload.doNotCreateStub !== null ? payload.doNotCreateStub : false,
+    })
+      .then((response) => Promise.resolve(response))
+      .catch((error) => Promise.reject(error));
+  },
 };
 
 const mutations = {};
@@ -120,5 +85,5 @@ export default {
   state,
   getters,
   mutations,
-  actions
+  actions,
 };
